// src/types/domain.ts
export enum Status {
  PENDING = 'Pendiente',
  READY = 'Listo',
  USING = 'En uso',
}
// ...
export type MachineType = 'Cabezal' | 'Casetera' | 'Freno';

export interface Banco {
  id: string;
  fecha_instalacion: string;
  lineas: number;
}

export interface InitialCabezalInput {
  id: string;
  estado?: Status;
  freno_id?: string;
  freno_estado?: Status;
  freno_fecha_inicio?: string;
}

export interface InitialCaseteraInput {
  id: number;
  estado?: Status;
}

export interface CreateEmpaqueDTO {
  id?: string;
  nombre: string;
  ubicacion: string;
  latitud: number;
  longitud: number;
  servicio: boolean;
  distancia: number;
  bancos: Banco[];
  cabezales?: InitialCabezalInput[];
  caseteras?: InitialCaseteraInput[];
}

export interface Empaque {
  id: string;
  nombre: string;
  ubicacion: string;
  latitud: number;
  longitud: number;
  servicio: boolean;
  distancia: number;
  bancos: Banco[];
  ultima_visita?: string;
}

export interface Tecnico {
  id: string;
  nombre: string;
  ultima_conexion: string;
  estado: StatusTecnico;
  cumpleanos: string;
  rol: 'tecnico' | 'admin';
  email: string;
  password?: string;
}

export interface Movimiento {
  id: string;
  machine_id: string;
  machine_type: MachineType;
  fecha: string;
  tecnico_nombre: string;
  motivo: string;
  origen: string;
  destino: string;
}

export interface Cabezal {
  id: string;
  tipo: 'Cabezal';
  estado: Status;
  ubicacion: string;
  freno_actual_id?: string;
  historial_movimientos: Movimiento[];
}

export interface Casetera {
  id: number;
  tipo: 'Casetera';
  estado: Status;
  ubicacion: string;
  historial_movimientos: Movimiento[];
}

export interface Freno {
  id: string;
  tipo: 'Freno';
  fecha_inicio: string;
  estado: Status;
  cabezal_id?: string;
  ubicacion: string;
  historial_movimientos: Movimiento[];
}